Refactor security (2/2).
Breaking one commit into two to try to preserve history.
